rename tokenclass to lm [pipe]
